Added WorkoutFeed component. Requests and renders workouts and sorts them by date and filters for unique entry. It uses infinite scroll component.
Refactored export code on all pages.

Add a '/api' as a prefix to all routes on the backend.

TODO
I need to make sure when registering a new user that
it checks the db for unique username.

Finish up the frontend display for the workoutsFeed.
Also adding functionality for it.