Added some simple backend santization for exercise weight. Fixed the decimal places to 2.
Changed the TokenExpirationHandler to sign out if
confirm is false, navigate to signin page if token
is expired to missing. Made the data base username
unique finally.

TODO
Add the ability to update and delete entries belonging
to the user.

Make ux for WorkoutFeed and workoutsCreation better.

Make sign out only able to be seen when signed in
and vice versa.